microservices/auth-service/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
PyJWT==2.6.0
Werkzeug==2.3.0
